Extract WorkflowFactory from the InitializeWorkflowActor
